slight changes to pkg/abi
In glibc, some structures are defined differently on different
platforms.
Such as: C.struct_stat
